+++ CHANGELOG.md
@@ -12,7 +12,17 @@ This project follows [Keep a Changelog](https://keepachangelog.com/) and [Semant
 ### Fixed
 -
 
-## [2.2.1] - 2025-08-30
+## [2.2.2] - 2025-09-07
+### Fixed
+- Fixed issue where it would give a false warning on secrets with process.env
+- Code cleanup.
+- exclude `.svelte-kit` from codebase scan by default.
+- Updated README
+
+### Changed
+- No breaking changes.
+
+## [2.2.1] - 2025-09-06
 ### Changed
 - tsconfig updates for improved type checking.
 - Updated codebase for new tsconfig rules